perf(ci): parallel InferenceTests, shorter ConcurrencyTests delays, split thinking E2E
- Pass --parallel to BaseChatInferenceTests CI step (~10-12 s saved per run) (#679)
- Drop SlowBackend token count from 20→5 in two ConcurrencyTests that assert
  concurrency behaviour, not output volume (~6.7 s saved locally) (#680)
- Split OllamaThinkingE2ETests into its own file so the three slow
  thinking tests (~180 s) can be targeted independently from the main
  Ollama suite (#614)
